libdrgn: get rid of compatible type requirement for {min,max}_iconst()
This has gotten in the way more than it has helped. I'll probably do the
same to min() and max() the next time they annoy me.

+++ libdrgn/hash_table.h
@@ -301,10 +301,7 @@ struct hash_table_iterator hash_table_first(struct hash_table *table);
 struct hash_table_iterator hash_table_next(struct hash_table_iterator it);
 #endif
 
-enum {
-	hash_table_chunk_alignment = max_iconst(alignof(max_align_t),
-						(size_t)16),
-};
+enum { hash_table_chunk_alignment = max_iconst(alignof(max_align_t), 16) };
 
 static inline size_t hash_table_probe_delta(struct hash_pair hp)
 {
